refactor tracker
